Move request generation to an extra module
